Split pass_stream_fusion.rs into directory with chain_detection, fusion_rules, lambda_composition, ir_transform
